requirements.txt
flask
flask-login
flask-sqlalchemy
python-dotenv
gunicorn
psycopg2-binary
Flask-Mail
Flask-Session
itsdangerous
Flask-Dance
feedparser
requests
google-genai>=0.3.0
google-cloud-bigquery>=3.25.0
yfinance
beautifulsoup4
openpyxl
pytest>=8.0.0

alembic
